+++ R/fitted.cmulti.R
@@ -1,7 +1,7 @@
 fitted.cmulti <-
 function(object, ...){
     type <- object$type
-    if (type == "mix")
+    if (type %in% c("mix", "fmix"))
         stop("fitted values are not available for finite mixture model")
     X <- model.matrix(object)
     poisson("log")$linkinv(drop(X %*% coef(object)))