+++ R/print.summary.cmulti.R
@@ -1,5 +1,5 @@
 print.summary.cmulti <-
-function (x, digits, ...) 
+function (x, digits, ...)
 {
     if (missing(digits))
         digits <- max(3, getOption("digits") - 3)
@@ -11,6 +11,8 @@ function (x, digits, ...)
         cat("Removal Sampling (homogeneous singing rate)\n")
     if (x$type == "mix")
         cat("Removal Sampling (heterogeneous singing rate)\n")
+    if (x$type == "fmix")
+        cat("Removal Sampling (heterogeneous singing rate)\n")
     cat(paste("Conditional Maximum Likelihood estimates\n\n", sep = ""))
     cat(paste("Coefficients:\n", sep = ""))
     printCoefmat(x$coefficients, digits = digits, signif.legend = FALSE)
@@ -18,7 +20,7 @@ function (x, digits, ...)
         if (getOption("show.signif.stars") & any(x$coefficients[,4] < 0.1))
             cat("---\nSignif. codes: ", "0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1", "\n")
     }
-    cat("\nLog-likelihood:", formatC(x$loglik, digits = digits), 
+    cat("\nLog-likelihood:", formatC(x$loglik, digits = digits),
         "\nBIC =", formatC(x$bic, digits = digits), "\n")
     cat("\n")
     invisible(x)