+++ NEWS.md
@@ -5,6 +5,7 @@
 * Denes et al. 2016 paper added to documentation.
 * `inst/ChangeLog` is removed and replaced by `NEWS.md`.
 * cmulti models now have a `predict` method, see examples.
+* fmix model type added to cmulti models.
 
 ## Version 0.4-0, March 2, 2016